% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_map_vec_to_matrix.R
\name{allele_map_vec_to_matrix}
\alias{allele_map_vec_to_matrix}
\title{Convert allele map vector to a binary matrix}
\usage{
allele_map_vec_to_matrix(allele_map)
}
\arguments{
\item{allele_map}{a vector of integers with length f equal to the number of founders}
}
\value{
a binary matrix, with dimension f by a, where a is the number of unique QTL alleles
}
\description{
Convert allele map vector to a binary matrix
}
