% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pvalue.R
\name{calc_pvalue}
\alias{calc_pvalue}
\title{Calculate a p-value from a permutation test}
\usage{
calc_pvalue(observed_lod, maxlods)
}
\arguments{
\item{observed_lod}{the test statistic, on the lod scale}

\item{maxlods}{a vector of max lods from a collection of permutations}
}
\value{
a permutation test p-value
}
\description{
Calculate a p-value from a permutation test
}
\details{
If the observed_lod is greater than all values in the vector maxlods, then a p-value of zero is returned. In practice, it means that you need more permutations.
}
