% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_timbr.R
\name{call_timbr}
\alias{call_timbr}
\title{Call timbr for use on cluster}
\usage{
call_timbr(mylist, traits_df, prior_M, genoprobs,
  A_matrix = mcv.data$prior.D$A)
}
\arguments{
\item{mylist}{a list containing two named components, probe and trait}

\item{traits_df}{a data.frame with one trait per column and column name corresponding to trait name}

\item{prior_M}{prior_M input for TIMBR::TIMBR()}

\item{genoprobs}{a n-long list containing, for each subject, a matrix of 36-state genotype probabilities at genome-wide collection of markers}

\item{addcovar}{additive covariates matrix to be appended to a column of 1's.}
}
\value{
output (a named list) of a single call to TIMBR
}
\description{
Call timbr for use on cluster
}
