% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_geno.R
\name{collapse_geno}
\alias{collapse_geno}
\title{Collapse a genotype probabilities array according to allelic series}
\usage{
collapse_geno(probs, allele_map)
}
\arguments{
\item{probs}{a genotype probabilities array (typically for one chromosome)}

\item{allele_map}{a numeric vector of integers mapping the founder alleles to QTL alleles}
}
\value{
collapsed genotypes array
}
\description{
Collapse a genotype probabilities array according to allelic series
}
